+++ plugins/ipmiinfo/js/ipmiinfo.js
@@ -132,6 +132,7 @@ function ipmiinfo_request() {
         ipmiinfo_populate(xml);
         if (ipmiinfo_show) {
             plugin_translate("ipmiinfo");
+            $("#Reload_ipmiinfoTable").attr("title",datetime());
             $("#Plugin_ipmiinfo").show();
         }
     }
@@ -148,6 +149,5 @@ $(document).ready(function ipmiinfo_buildpage() {
 
     $("#Reload_ipmiinfoTable").click(function ipmiinfo_reload(id) {
         ipmiinfo_request();
-        $("#Reload_ipmiinfoTable").attr("title",datetime());
     });
 });