+++ plugins/psstatus/js/psstatus.js
@@ -103,6 +103,7 @@ function psstatus_request() {
             psstatus_populate(xml);
             if (psstatus_show) {
                 plugin_translate("PSStatus");
+                $("#Reload_PSStatusTable").attr("title",datetime());
                 $("#Plugin_PSStatus").show();
             }
         }
@@ -119,6 +120,5 @@ $(document).ready(function psstatus_buildpage() {
 
     $("#Reload_PSStatusTable").click(function psstatus_reload(id) {
         psstatus_request();
-        $("#Reload_PSStatusTable").attr("title",datetime());
     });
 });