+++ plugins/smart/js/smart.js
@@ -143,6 +143,7 @@ function smart_request() {
             smart_populate(xml);
             if (smart_show) {
                 plugin_translate("SMART");
+                $("#Reload_SMARTTable").attr("title",datetime());
                 $("#Plugin_SMART").show();
             }
         }
@@ -159,6 +160,5 @@ $(document).ready(function smart_buildpage() {
 
     $("#Reload_SMARTTable").click(function smart_reload(id) {
         smart_request();
-        $("#Reload_SMARTTable").attr("title",datetime());
     });
 });