+++ plugins/snmppinfo/js/snmppinfo.js
@@ -120,6 +120,7 @@ function snmppinfo_request() {
             snmppinfo_buildTable(xml);
             if (snmppinfo_show) {
                 plugin_translate("SNMPPInfo");
+                $("#Reload_SNMPPInfoTable").attr("title",datetime());
                 $("#Plugin_SNMPPInfo").show();
             }
         }
@@ -134,6 +135,5 @@ $(document).ready(function snmppinfo_buildpage() {
 
     $("#Reload_SNMPPInfoTable").click(function snmppinfo_reload(id) {
         snmppinfo_request();
-        $("#Reload_SNMPPInfoTable").attr("title",datetime());
     });
 });