+++ plugins/uprecords/js/uprecords.js
@@ -95,6 +95,7 @@ function uprecords_request() {
         uprecords_populate(xml);
         if (uprecords_show) {
             plugin_translate("uprecords");
+            $("#Reload_uprecordsTable").attr("title",datetime());
             $("#Plugin_uprecords").show();
         }
     }
@@ -111,6 +112,5 @@ $(document).ready(function uprecords_buildpage() {
 
     $("#Reload_uprecordsTable").click(function uprecords_reload(id) {
         uprecords_request();
-        $("#Reload_uprecordsTable").attr("title",datetime());
     });
 });