+++ plugins/uprecords/class.uprecords.inc.php
@@ -78,14 +78,16 @@ class uprecords extends PSI_Plugin
                 $lines = "";
                 $oldtz=getenv("TZ");
                 putenv("TZ=GMT");
-				if(PSI_PLUGIN_UPRECORDS_MAX <= 10 && PSI_PLUGIN_UPRECORDS_MAX > 0 )
-				{
-					$options=" -m".PSI_PLUGIN_UPRECORDS_MAX;
-				}
-				else
-				{
-					$options="";
-				}
+                $options = "";
+                if (defined('PSI_PLUGIN_UPRECORDS_MAX_ENTRIES')) {
+                    if (PSI_PLUGIN_UPRECORDS_MAX_ENTRIES === false) {
+                        $options=" -m 0";
+                    } elseif (PSI_PLUGIN_UPRECORDS_MAX_ENTRIES === true) {
+                        $options=" -m 1";
+                    } elseif (PSI_PLUGIN_UPRECORDS_MAX_ENTRIES > 1) {
+                        $options=" -m ".PSI_PLUGIN_UPRECORDS_MAX_ENTRIES;
+                    }
+                }
                 if (CommonFunctions::executeProgram('uprecords', '-a -w'.$options, $lines) && !empty($lines))
                     $this->_lines = preg_split("/\n/", $lines, -1, PREG_SPLIT_NO_EMPTY);
                 putenv("TZ=".$oldtz);