+++ CHANGELOG.md
@@ -7,7 +7,7 @@ phpSysInfo 3.2.4
 ----------------
  - [ADD] Solus to detected distros
  - [ADD] Uprecords plugin - MAX_ENTRIES option - maximum number of entries to show
- 
+
  - [UPD] Refreshing status of all plugins for dynamic mode
 
  - [NEW] language and template selection also for frontend "bootstrap"